+++ src/rs/section.rs
@@ -31,6 +31,8 @@ pub struct Section {
 
     pub align: Option<u64>,
 
+    pub is_fill: bool,
+
     pub symbols: Vec<symbol::Symbol>,
 }
 
@@ -43,6 +45,18 @@ impl Section {
         vrom: Option<u64>,
         align: Option<u64>,
     ) -> Self {
+        Self::new_impl(filepath, vram, size, section_type, vrom, align, false)
+    }
+
+    pub(crate) fn new_impl(
+        filepath: PathBuf,
+        vram: u64,
+        size: u64,
+        section_type: &str,
+        vrom: Option<u64>,
+        align: Option<u64>,
+        is_fill: bool,
+    ) -> Self {
         Self {
             filepath,
             vram,
@@ -50,6 +64,7 @@ impl Section {
             section_type: section_type.into(),
             vrom,
             align,
+            is_fill,
             symbols: Vec::new(),
         }
     }
@@ -301,7 +316,7 @@ impl Section {
 }
 
 impl Section {
-    pub fn new_default(
+    pub(crate) fn new_default(
         filepath: std::path::PathBuf,
         vram: u64,
         size: u64,
@@ -314,23 +329,12 @@ impl Section {
             section_type: section_type.into(),
             vrom: None,
             align: None,
+            is_fill: false,
             symbols: Vec::new(),
         }
     }
 
-    pub fn clone_no_symbollist(&self) -> Self {
-        Section {
-            filepath: self.filepath.clone(),
-            vram: self.vram,
-            size: self.size,
-            section_type: self.section_type.clone(),
-            vrom: self.vrom,
-            align: self.align,
-            symbols: Vec::new(),
-        }
-    }
-
-    pub fn new_placeholder() -> Self {
+    pub(crate) fn new_placeholder() -> Self {
         Self {
             filepath: "".into(),
             vram: 0,
@@ -338,6 +342,25 @@ impl Section {
             section_type: "".into(),
             vrom: None,
             align: None,
+            is_fill: false,
+            symbols: Vec::new(),
+        }
+    }
+
+    pub(crate) fn new_fill(
+        filepath: std::path::PathBuf,
+        vram: u64,
+        size: u64,
+        section_type: &str,
+    ) -> Self {
+        Self {
+            filepath,
+            vram,
+            size,
+            section_type: section_type.into(),
+            vrom: None,
+            align: None,
+            is_fill: true,
             symbols: Vec::new(),
         }
     }
@@ -436,7 +459,7 @@ pub(crate) mod python_bindings {
     #[pymethods]
     impl Section {
         #[new]
-        #[pyo3(signature = (filepath, vram, size, section_type, vrom=None, align=None))]
+        #[pyo3(signature = (filepath, vram, size, section_type, vrom=None, align=None, is_fill=false))]
         fn py_new(
             filepath: PathBuf,
             vram: u64,
@@ -444,8 +467,9 @@ pub(crate) mod python_bindings {
             section_type: &str,
             vrom: Option<u64>,
             align: Option<u64>,
+            is_fill: bool,
         ) -> Self {
-            Self::new(filepath, vram, size, section_type, vrom, align)
+            Self::new_impl(filepath, vram, size, section_type, vrom, align, is_fill)
         }
 
         /* Getters and setters */
@@ -520,6 +544,17 @@ pub(crate) mod python_bindings {
         #[setter]
         fn set_align(&mut self, value: Option<u64>) -> PyResult<()> {
             self.align = value;
+            Ok(())
+        }
+
+        #[getter]
+        fn get_isFill(&self) -> PyResult<bool> {
+            Ok(self.is_fill)
+        }
+
+        #[setter]
+        fn set_isFill(&mut self, value: bool) -> PyResult<()> {
+            self.is_fill = value;
             Ok(())
         }